+++ tests/testthat/test-genio.R
@@ -715,11 +715,11 @@ test_that("make_bim works", {
 
 test_that("require_files_plink works", {
     # these should all just work (pass existing files)
-    require_files_plink('dummy-33-101-0.1')
-    require_files_plink('dummy-4-10-0.1')
-    require_files_plink('dummy-5-10-0.1')
-    require_files_plink('dummy-6-10-0.1')
-    require_files_plink('dummy-7-10-0.1')
+    expect_silent( require_files_plink('dummy-33-101-0.1') )
+    expect_silent( require_files_plink('dummy-4-10-0.1') )
+    expect_silent( require_files_plink('dummy-5-10-0.1') )
+    expect_silent( require_files_plink('dummy-6-10-0.1') )
+    expect_silent( require_files_plink('dummy-7-10-0.1') )
     # try something that doesn't exist, expect to fail
     expect_error( require_files_plink('file-that-does-not-exist') )
 })