+++ NEWS.md
@@ -87,5 +87,5 @@
 
 # 2020-06-16 - genio 1.0.13.9000
 
-* Added `read_grm` to read GCTA's binary genetic relatedness matrix (GRM) format.
-
+* Added `read_grm` and `write_grm` to read and write GCTA's binary genetic relatedness matrix (GRM) format.
+* Also added auxiliary functions `require_files_grm` and `delete_files_grm`.